pyproject.toml
[tool.poetry]
name = "p1-influx-db"
version = "1.0.0"
description = "Parsing P1 serial to Influx DB"
authors = ["Yorick Fredrix <yorick.fredrix@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
dsmr-parser = "^1.3.1"
loguru = "^0.7.2"
# A list of all of the optional dependencies, some of which are included in the
# below `extras`. They can be opted into by apps.
influxdb-client = {extras = ["async"], version = "^1.39.0"}
paho-mqtt = "^2.1.0"

[tool.poetry.extras]
influxdb = ["influxdb-client"]
mqtt = ["paho-mqtt"]
publishmethods = ["influxdb-client", "paho-mqtt"]

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest = "^7.4.4"
python-semantic-release = "^9.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = []

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = true

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["build", "chore", "ci", "docs", "feat", "fix", "perf", "style", "refactor", "test"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

